+++ test/sql/functions.sql
@@ -22,6 +22,10 @@ SELECT cosine_distance('[1,1]', '[1,1]');
 SELECT cosine_distance('[1,1]', '[-1,-1]');
 SELECT cosine_distance('[1,2]', '[3]');
 
+SELECT l1_distance('[0,0]', '[3,4]');
+SELECT l1_distance('[0,0]', '[0,1]');
+SELECT l1_distance('[1,2]', '[3]');
+
 SELECT avg(v) FROM unnest(ARRAY['[1,2,3]'::vector, '[3,5,7]']) v;
 SELECT avg(v) FROM unnest(ARRAY['[1,2,3]'::vector, '[3,5,7]', NULL]) v;
 SELECT avg(v) FROM unnest(ARRAY[]::vector[]) v;